MeshImporterからMeshContextを分離
